+++ tests/corpus/interfaces_readers_test.py
@@ -16,7 +16,7 @@ def test_text_tokenizer_interface():
     # assert [('document_1.txt', ['a', 'b', 'c']), ('document_2.txt', ['e', 'f', 'g'])] == [ x for x in corpus ]
 
 
-def test_dataframe_text_tokenizer_interface():
+def test_pandas_reader_interface():
 
     assert issubclass(readers.PandasCorpusReader, ICorpusReader)
